move functions to helpers folder
