+++ helpers/textCleaner.js
