+++ helpers/typeCheck.js
