+++ helpers/validateInput.js
